deploy-requirements.txt
apscheduler==3.6.3
attrs==21.2.0
backports.zoneinfo==0.2.1; python_version < '3.9'
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.6; python_version >= '3'
idna==3.2; python_version >= '3'
iniconfig==1.1.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==3.0
urllib3==1.26.7
